+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/VirtualizedFileSystemTest.java
@@ -690,15 +690,19 @@ public class VirtualizedFileSystemTest {
     }
 
     @Test
-    public void testToUriRelative() {
+    public void testToRelativeUri() {
         final Context ctx = cfg.getContext();
         final Path userDir = cfg.getUserDir();
         languageAction = (Env env) -> {
-            TruffleFile file = env.getTruffleFile(FILE_EXISTING);
-            URI uri = file.toUri(false);
+            TruffleFile relativeFile = env.getTruffleFile(FILE_EXISTING);
+            URI uri = relativeFile.toRelativeUri();
             Assert.assertFalse(uri.isAbsolute());
             URI expectedUri = userDir.toUri().relativize(userDir.resolve(FILE_EXISTING).toUri());
             Assert.assertEquals(cfg.formatErrorMessage("Relative URI"), expectedUri, uri);
+            final TruffleFile absoluteFile = env.getTruffleFile("/").resolve(FOLDER_EXISTING).resolve(FILE_EXISTING);
+            uri = absoluteFile.toUri();
+            Assert.assertTrue(uri.isAbsolute());
+            Assert.assertEquals(cfg.formatErrorMessage("Absolute URI"), Paths.get("/").resolve(FOLDER_EXISTING).resolve(FILE_EXISTING).toUri(), uri);
         };
         ctx.eval(LANGAUGE_ID, "");
     }