+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/TruffleFile.java
@@ -320,35 +320,32 @@ public final class TruffleFile {
      */
     @TruffleBoundary
     public URI toUri() {
-        return toUri(true);
+        try {
+            final Path absolutePath = path.isAbsolute() ? path : toAbsolutePathImpl()[0];
+            return absolutePath.toUri();
+        } catch (SecurityException se) {
+            throw se;
+        } catch (Throwable t) {
+            throw wrapHostException(t);
+        }
     }
 
     /**
-     * Returns possibly relative {@link URI} representation of this {@link TruffleFile}. When
-     * invoked with {@code absolute} parameter set to {@code true} this method always creates an
-     * absolute {@link URI}. For {@code absolute} set to {@code false} this method creates absolute
-     * {@link URI} for {@link #isAbsolute() absolute} {@link TruffleFile}s and relative {@link URI}
-     * for relative ones.
+     * Returns a relative {@link URI} representation of non absolute {@link TruffleFile}. If this
+     * {@link TruffleFile} is relative it returns a relative {@link URI}. For an
+     * {@link #isAbsolute() absolute} {@link TruffleFile} it returns an absolute {@link URI}.
      *
-     * @param absolute if {@code true} the returned {@link URI} is absolute even for relative
-     *            {@link TruffleFile}
      * @return the {@link URI} representing the {@link TruffleFile}
-     * @throws SecurityException if an absolute {@link URI} is required and the {@link FileSystem}
-     *             denied a resolution of an absolute path
      * @since 1.0
      */
     @TruffleBoundary
-    public URI toUri(boolean absolute) {
+    public URI toRelativeUri() {
+        if (isAbsolute()) {
+            return toUri();
+        }
         try {
-            if (absolute || path.isAbsolute()) {
-                final Path absolutePath = path.isAbsolute() ? path : toAbsolutePathImpl()[0];
-                return absolutePath.toUri();
-            } else {
-                String strPath = "/".equals(path.getFileSystem().getSeparator()) ? path.toString() : path.toString().replace(path.getFileSystem().getSeparator(), "/");
-                return new URI(null, null, strPath, null);
-            }
-        } catch (SecurityException se) {
-            throw se;
+            String strPath = "/".equals(path.getFileSystem().getSeparator()) ? path.toString() : path.toString().replace(path.getFileSystem().getSeparator(), "/");
+            return new URI(null, null, strPath, null);
         } catch (Throwable t) {
             throw wrapHostException(t);
         }