+++ truffle/CHANGELOG.md
@@ -7,7 +7,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Removed deprecated `TruffleLanguage.Env.newSourceBuilder`.
 * Added `TruffleLanguage.Env.isPreInitialization` method to determine whether the context is being pre-initialized.
 * Added `ArrayUtils` API providing additional array and/or string operations that may be intrinsified by the compiler.
-* Added a possibility to obtain a [relative URI](http://www.graalvm.org/truffle/javadoc/file:///Users/tom/Projects/graal/tzezula/graal/truffle/javadoc/com/oracle/truffle/api/TruffleFile.html#toUri-boolean-) for a relative `TruffleFile`.
+* Added a possibility to obtain a [relative URI](http://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/TruffleFile.html#toRelativeUri--) for a relative `TruffleFile`.
 
 ## Version 1.0.0 RC9